+++ tests/unit/features/overlay/test_parts.py
@@ -48,6 +48,7 @@ class TestPartSpecs:
             "disable-parallel": True,
             "after": ["bar"],
             "overlay-packages": ["overlay-pkg1", "overlay-pkg2"],
+            "overlay-recommended-packages": [],
             "stage-snaps": ["stage-snap1", "stage-snap2"],
             "stage-packages": ["stage-pkg1", "stage-pkg2"],
             "build-snaps": ["build-snap1", "build-snap2"],
@@ -128,20 +129,30 @@ class TestPartSpecs:
         assert spec.stage_packages == package_list
 
     @pytest.mark.parametrize(
-        ("packages", "overlay_script", "override_script", "files", "result"),
+        (
+            "packages",
+            "rec_packages",
+            "overlay_script",
+            "override_script",
+            "files",
+            "result",
+        ),
         [
-            ([], None, None, ["*"], False),
-            (["pkg"], None, None, ["*"], True),
-            ([], "ls", None, ["*"], True),
-            ([], None, "ls", ["*"], True),
-            ([], None, None, ["-usr/share"], True),
+            ([], [], None, None, ["*"], False),
+            (["pkg"], [], None, None, ["*"], True),
+            ([], [], "ls", None, ["*"], True),
+            ([], [], None, "ls", ["*"], True),
+            ([], [], None, None, ["-usr/share"], True),
+            ([], ["pkg"], None, None, ["*"], True),
+            (["pkg1"], ["pkg2"], None, None, ["*"], True),
         ],
     )
     def test_spec_has_overlay(
-        self, packages, overlay_script, override_script, files, result
+        self, packages, rec_packages, overlay_script, override_script, files, result
     ):
         data = {
             "overlay-packages": packages,
+            "overlay-recommended-packages": rec_packages,
             "overlay-script": overlay_script,
             "override-overlay": override_script,
             "overlay": files,